requirements.txt
South>=0.7.3,<0.8

# Developing
Werkzeug
django-debug-toolbar
django-extensions

# Testing
mock>=1.0.1,<1.1
django-nose>=1.1,<1.2
spec
coverage>=3.6,<3.7
django-dynamic-fixture
freezegun
